+++ packages/nesp_sdk_flutter_ui/lib/painting.dart
@@ -3,3 +3,4 @@ library;
 export 'src/painting/box_border.dart';
 export 'src/painting/rounded_rectangle_border.dart';
 export 'src/painting/tooltip_shape_border.dart';
+export 'src/painting/image.dart';