+++ packages/nesp_sdk_flutter_ui/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.1-dev.2
+
+* Add ListItem, ListGroup, DashLine and ChangeNotifierViewModelContent.
+* Updates ViewModel to support visibility changes.
+* Add Image to support save image listen image load from file.
+
 ## 0.0.1-dev.1
 
 * The Initial version